+++ csa-runner-app/src/main/java/com/farao_community/farao/swe_csa/app/dichotomy/SweCsaRaoValidator.java
@@ -1,5 +1,16 @@
 package com.farao_community.farao.swe_csa.app.dichotomy;
 
+/*
+ * Copyright (c) 2024, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+/**
+ * @author Jean-Pierre Arnould {@literal <jean-pierre.arnould at rte-france.com>}
+ */
+
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_io_json.JsonImport;
 import com.farao_community.farao.data.rao_result_api.RaoResult;
